script: three casts made a 32-bit operand 16 bits wide, and the corpus was at 49% of it
`op::closure` names its target with `with_bx`, which takes a `std::uint32_t` and
splits it across two operand fields. Three of the four sites that emit it cast
the function index to `std::uint16_t` first. That is not a bound - nothing
checked anything - it is a WRAP, and the compiler silently built a closure over
a different function.

Measured, before the casts went:

  70,001 functions, calling function 69,999  ->  ran function 4,463
                                                 69,999 - 65,536, exactly

No error from the compiler, none from the VM, none from the image validator. The
program compiled, loaded, ran, and returned the wrong number. Babylon, which is
vendored in this repository, is 31,905 functions - 49% of a ceiling the
instruction encoding never had.

Deleting the three casts is the whole fix: `statements.cpp` for a function
declaration, `expressions.cpp` for an arrow, `classes.cpp` for a field
initialiser. `classes.cpp`'s constructor site never had one, which is why this
was survivable at all. Verified at 70,001 and 140,001 functions - 34 ms to
compile the larger - both calling the function they name.

THE TEST PINS THE INSTRUCTION, NOT THE OUTPUT. Running the program would prove
the same thing while saying less; the assertion is that every `op::closure`
names a function that exists and that the highest index emitted is past 65,535.
It fails on restoring any one of the three casts, which is how it was checked.
It also has to build its 70,001 functions under ONE name, because the compiler
refuses a program with more than 65,536 distinct property names - a real guard,
for the operand that selects one - and it would otherwise fire first and prove
nothing about this.

AND THE IMAGE'S OWN 65,535 REFUSAL GOES WITH IT. It existed to avoid freezing
the compiler's mistake into a file, and said so. What replaces it is not a bigger
constant: the loader now checks the function count against the bytes remaining,
at 49 bytes minimum per function, because `resize` on a number out of a file is
the reserve-the-world bug that `read_pool` was already fixed for once.

97 of 97 on the devbox.

+++ ctbrowser/lib/Script/compile/classes.cpp
@@ -153,7 +153,7 @@ void compiler_impl::compile_class(const vp::node & n, std::uint16_t dst, bool as
     if (!instance_fields.empty()) {
         const std::uint32_t fields = compile_field_initialiser(instance_fields);
         const std::uint16_t init = alloc_reg();
-        proto().emit(instruction::with_bx(op::closure, init, static_cast<std::uint16_t>(fields)));
+        proto().emit(instruction::with_bx(op::closure, init, fields));
         proto().emit(instruction{op::set_prop, dst, name_operand("__fields"), init});
     }
